+++ glue/viewers/matplotlib/qt/data_viewer.py
@@ -8,11 +8,11 @@ from glue.viewers.common.viz_client import init_mpl, update_appearance_from_sett
 from glue.external.echo import delay_callback
 from glue.utils import defer_draw, mpl_to_datetime64
 from glue.utils.decorators import avoid_circular
-from glue.viewers.matplotlib.qt.toolbar import MatplotlibViewerToolbar
 from glue.viewers.matplotlib.state import MatplotlibDataViewerState
-from glue.viewers.image.layer_artist import ImageSubsetLayerArtist
-from glue.core.edit_subset_mode import EditSubsetMode
 from glue.core.command import ApplySubsetState
+
+# The following import is required to register the viewer tools
+from glue.viewers.matplotlib.qt import toolbar  # noqa
 
 __all__ = ['MatplotlibDataViewer']
 
@@ -46,8 +46,10 @@ plt.close(fig)
 
 class MatplotlibDataViewer(DataViewerWithState):
 
-    _toolbar_cls = MatplotlibViewerToolbar
     _state_cls = MatplotlibDataViewerState
+
+    tools = ['mpl:home', 'mpl:pan', 'mpl:zoom']
+    subtools = {'save': ['mpl:save']}
 
     def __init__(self, session, parent=None, wcs=None, state=None):